+++ index.js
@@ -4,3 +4,4 @@ require('./lib/apiVerificationSteps.js');
 require('./lib/parameterTypes.js');
 require('./lib/apiConstructionSteps.js');
 require('./lib/websocket.js');
+require('./lib/grpc.js');